\name{fun_3}
\alias{fun_3}
\title{Title}
\description{Description.}
\usage{
fun_3(arg1, arg2, arg3)
}
\arguments{
  \item{arg1}{arg1.}
  \item{arg2}{arg2.}
  \item{arg3}{arg3).
  }
  \item{...}{no such argument!}
}
\details{
  Details.

  a doi: \doi{10.18637/jss.v040.i03}

  another doi: \doi{10.1016/S0024-3795(01)00475-X}
}

\RdOpts{stage=build}
\references{
 a doi: \doi{10.18637/jss.v040.i03}
}
\RdOpts{stage=install}

\examples{
fun_3()
}
